move runCode
