+++ lib/contracts/deploy.js
@@ -2,7 +2,7 @@ let async = require('async');
 //require("../utils/debug_util.js")(__filename, async);
 let utils = require('../utils/utils.js');
 
-let RunCode = require('../core/runCode.js');
+let RunCode = require('../coderunner/runCode.js');
 
 class Deploy {
   constructor(options) {
