+++ lib/core/engine.js
@@ -12,6 +12,7 @@ const ServicesMonitor = require('./services_monitor.js');
 const Pipeline = require('../pipeline/pipeline.js');
 const Watch = require('../pipeline/watch.js');
 const LibraryManager = require('../versions/library_manager.js');
+const CodeRunner = require('../coderunner/codeRunner.js');
 
 class Engine {
   constructor(options) {
@@ -161,6 +162,12 @@ class Engine {
 
   codeGeneratorService(_options) {
     let self = this;
+
+    this.codeRunner = new CodeRunner({
+      plugins: self.plugins,
+      events: self.events,
+      logger: self.logger
+    });
 
     this.codeGenerator = new CodeGenerator({
       blockchainConfig: self.config.blockchainConfig,
