+++ lib/dashboard/console.js
@@ -1,5 +1,5 @@
 let utils = require('../utils/utils.js');
-let RunCode = require('../core/runCode.js');
+let RunCode = require('../coderunner/runCode.js');
 
 class Console {
   constructor(options) {
