+++ lib/modules/specialconfigs/index.js
@@ -1,4 +1,4 @@
-let RunCode = require('../../core/runCode.js');
+let RunCode = require('../../coderunner/runCode.js');
 const stringReplaceAsync = require('string-replace-async');
 const async = require('async');
