+++ lib/tests/test.js
@@ -2,7 +2,7 @@ var async = require('async');
 //require("../utils/debug_util.js")(__filename, async);
 var Web3 = require('web3');
 var Engine = require('../core/engine.js');
-var RunCode = require('../core/runCode.js');
+var RunCode = require('../coderunner/runCode.js');
 var TestLogger = require('./test_logger.js');
 
 var getSimulator = function() {
